Move code for the activist portal into new "canvass" feature folder.
