+++ src/features/canvass/components/EncouragingSparkle.tsx
