+++ src/features/canvass/utils/getCrosshairPositionOnMap.tsx
