+++ src/features/canvass/utils/getVisitState.spec.ts
