+++ src/features/canvass/utils/isPointInsidePolygon.ts
