+++ src/features/canvass/utils/markerIcon.tsx
