+++ src/features/home/hooks/useMyActivities.ts
@@ -1,4 +1,4 @@
-import useMyAreaAssignments from 'features/areaAssignments/hooks/useMyAreaAssignments';
+import useMyAreaAssignments from 'features/canvass/hooks/useMyAreaAssignments';
 import useMyCallAssignments from 'features/callAssignments/hooks/useMyCallAssignments';
 import useMyEvents from 'features/events/hooks/useMyEvents';
 import { MyActivity } from '../types';
